requirements.txt
# Data Version Control & Pipeline Orchestration
dvc[s3]
pyyaml

# Core ML Libraries
pandas
numpy
scikit-learn
nltk

# MLflow & Experiment Tracking
mlflow
dagshub

# API Framework
fastapi
uvicorn
python-multipart
pydantic

# Frontend
streamlit
requests

# Data Visualization & Analysis
seaborn
matplotlib
jupyter
ipywidgets
tqdm

# Optional: Deep Learning (commented out for cpu-only)
# torch --index-url https://download.pytorch.org/whl/cpu
# transformers
# huggingface_hub

# Development & Utilities
python-dotenv
click

